+++ packages/manager/apps/carrier-sip/src/cdr/routing.js
@@ -1,0 +1,6 @@
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('app.cdrs', {
+    url: '/cdr',
+    component: 'carrierSipCdr',
+  });
+};