+++ packages/manager/apps/dedicated/client/app/account/contacts/service/contacts-service.component.js
@@ -1,0 +1,12 @@
+import controller from './user-contacts-service.controller';
+import template from './user-contacts-service.html';
+
+export default {
+  bindings: {
+    editContacts: '<',
+    getServiceInfos: '<',
+    services: '<',
+  },
+  controller,
+  template,
+};