+++ packages/manager/apps/dedicated/client/app/billing/autoRenew/actions/terminate-enterprise-cloud-database/terminate-enterprise-cloud-database.controller.js
@@ -1,0 +1,19 @@
+export default class {
+  /* @ngInject */
+  constructor($translate) {
+    this.$translate = $translate;
+  }
+
+  onSuccess() {
+    this.goBack(
+      this.$translate.instant('autorenew_enterprise_cloud_database_close_service_success'),
+    );
+  }
+
+  onError(error) {
+    this.goBack(
+      this.$translate.instant('autorenew_enterprise_cloud_database_close_service_error', _.get(error, 'data.message')),
+      'danger',
+    );
+  }
+}