+++ packages/manager/apps/dedicated/client/app/billing/autoRenew/disable/disable.controller.js
@@ -6,8 +6,8 @@ export default class {
     this.$translate = $translate;
   }
 
-  update() {
-    return this.updateRenew()
+  update(services) {
+    return this.updateRenew(services)
       .then(() => this.goBack(
         this.$translate.instant('billing_autorenew_disable_success'),
       ))