+++ packages/manager/apps/dedicated/client/app/components/no-default-means-of-payment/no-default-means-of-payment.component.js
@@ -1,0 +1,9 @@
+import template from './no-default-means-of-payment.html';
+
+export default {
+  bindings: {
+    onGoToMeansOfPayment: '&',
+  },
+  name: 'ovhManagerNoDefaultMeansOfPayment',
+  template,
+};