+++ packages/manager/apps/dedicated/client/app/components/stepper/step/step.header.js
@@ -1,0 +1,9 @@
+export default class {
+  constructor(
+    displayValue,
+    isVisible = true,
+  ) {
+    this.displayValue = displayValue;
+    this.isVisible = isVisible;
+  }
+}