+++ packages/manager/apps/dedicated/client/app/dedicated/server/interfaces/ola-activation/ola-activation.component.js
@@ -1,0 +1,14 @@
+import controller from './ola-activation.controller';
+import template from './ola-activation.html';
+
+export default {
+  bindings: {
+    goBack: '<',
+    alertError: '<',
+    atTrack: '<',
+    serverName: '<',
+    user: '<',
+  },
+  controller,
+  template,
+};