+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/dashboard/tiles/options/models/user/index.js
@@ -1,0 +1,12 @@
+import {
+  name as serviceName,
+  OptionsUserService,
+} from './user.service';
+
+const moduleName = 'ovhManagerPccDashboardOptionsUser';
+
+angular
+  .module(moduleName, [])
+  .service(serviceName, OptionsUserService);
+
+export default moduleName;