+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/dashboard/tiles/service-management/index.js
@@ -1,0 +1,14 @@
+import component from './service-management.component';
+
+const moduleName = 'ovhManagerPccDashboardServiceManagement';
+
+angular
+  .module(moduleName, [
+    'oui',
+    'pascalprecht.translate',
+    'ui.router',
+  ])
+  .component(component.name, component)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;