+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/mailing-list/subscribe/dedicatedCloud-mailing-list-subscribe.controller.js
@@ -33,9 +33,9 @@ angular.module('App').controller('DedicatedCloudMailingCtrl', class DedicatedClo
 
     return this.dedicatedCloudMailingList.postMailingList(this.model.email, this.pccMl).then(() => this.Alerter.success(this.$translate.instant('dedicatedCloud_subscribe_mailing_step2_success', {
       t0: this.model.email,
-    }), 'dedicatedCloud')).catch(error => this.Alerter.error([this.$translate.instant('dedicatedCloud_subscribe_mailing_step2_error', {
+    }))).catch(error => this.Alerter.error([this.$translate.instant('dedicatedCloud_subscribe_mailing_step2_error', {
       t0: this.model.email,
-    }), get(error, 'message')].join('. '), 'dedicatedCloud')).finally(() => {
+    }), get(error, 'message')].join('. '))).finally(() => {
       this.onWidzardCancel();
       this.loading.subscribe = false;
     });