+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/service-pack/upgrade/steps/selection/selection.component.js
@@ -1,0 +1,22 @@
+import controller from './selection.controller';
+import template from './selection.html';
+
+export default {
+  bindings: {
+    activationType: '<',
+    currentUser: '<',
+    currentService: '<',
+    hasDefaultMeansOfPayment: '<',
+    header: '@',
+    orderableServicePacks: '<',
+    servicePacksWithPrices: '<',
+    servicePackToOrder: '<?',
+    subHeader: '@',
+  },
+  controller,
+  name: 'ovhManagerPccServicePackUpgradeSelection',
+  require: {
+    stepper: '^ovhManagerComponentStepper',
+  },
+  template,
+};