+++ packages/manager/apps/dedicated/client/app/dedicatedCloud/terminate/confirm/dedicatedCloud-terminate-confirm.controller.js
@@ -48,12 +48,12 @@ angular.module('App').controller('DedicatedCloudConfirmTerminateCtrl', class Ded
       reason: this.model.reason,
       token: this.$stateParams.token,
     }).$promise.then(() => {
-      this.Alerter.success(this.$translate.instant('dedicatedCloud_confirm_close_success'), 'dedicatedCloud_alert');
+      this.Alerter.success(this.$translate.instant('dedicatedCloud_confirm_close_success'));
     }).catch((error) => {
       this.Alerter.alertFromSWS(this.$translate.instant('dedicatedCloud_confirm_close_error'), {
         message: get(error, 'message'),
         type: 'ERROR',
-      }, 'dedicatedCloud_alert');
+      });
     }).finally(() => {
       this.loading.confirm = false;
       this.$state.go('^');