+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/comfortExchange/pack-xdsl-access-comfort-exchange.routing.js
@@ -1,0 +1,17 @@
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider
+    .state('telecom.pack.xdsl.access-modem-exchange', {
+      url: '/comfortExchange',
+      views: {
+        'accessView@telecom.pack.xdsl': 'xdslAccessComfortExchange',
+      },
+      resolve: {
+        xdslId: /* @ngInject */ $transition$ => $transition$.params().serviceName,
+        openedRMAs: /* @ngInject */ (
+          XdslAccessComfortExchangeService,
+          xdslId,
+        ) => XdslAccessComfortExchangeService.getOpenedRMAs(xdslId),
+      },
+      translations: { value: ['..', '.'], format: 'json' },
+    });
+};