+++ packages/manager/modules/enterprise-cloud-database/src/create/choose-database/choose-database.controller.js
@@ -1,7 +1,6 @@
 export default class {
+  /* @ngInject */
   constructor($timeout) {
-    'ngInject';
-
     this.$timeout = $timeout;
   }