+++ packages/manager/modules/enterprise-cloud-database/CHANGELOG.md
@@ -1,3 +1,23 @@
+## [0.1.2](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-enterprise-cloud-database@0.1.1...@ovh-ux/manager-enterprise-cloud-database@0.1.2) (2019-10-17)
+
+
+### Bug Fixes
+
+* **deps:** upgrade @ovh-ux/component-rollup-config to v7.0.0 ([#1469](https://github.com/ovh-ux/manager/issues/1469)) ([bbc8794](https://github.com/ovh-ux/manager/commit/bbc8794))
+* **i18n:** add missing translations [CDS 104] ([da22965](https://github.com/ovh-ux/manager/commit/da22965))
+
+
+
+## [0.1.1](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-enterprise-cloud-database@0.1.0...@ovh-ux/manager-enterprise-cloud-database@0.1.1) (2019-10-15)
+
+
+### Bug Fixes
+
+* **enterprise-cloud-database:** cleanup and mock data removal ([#1434](https://github.com/ovh-ux/manager/issues/1434)) ([5eca0e8](https://github.com/ovh-ux/manager/commit/5eca0e8))
+* **i18n:** add missing translations [CDS 98] ([e180e8a](https://github.com/ovh-ux/manager/commit/e180e8a))
+
+
+
 # [0.1.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-enterprise-cloud-database@0.0.0...@ovh-ux/manager-enterprise-cloud-database@0.1.0) (2019-10-07)