+++ packages/manager/modules/telecom-dashboard/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [4.4.1](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-telecom-dashboard@4.4.0...@ovh-ux/manager-telecom-dashboard@4.4.1) (2019-10-17)
+
+
+### Bug Fixes
+
+* **deps:** upgrade @ovh-ux/component-rollup-config to v7.0.0 ([#1469](https://github.com/ovh-ux/manager/issues/1469)) ([bbc8794](https://github.com/ovh-ux/manager/commit/bbc8794))
+
+
+
 # [4.4.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-telecom-dashboard@4.3.2...@ovh-ux/manager-telecom-dashboard@4.4.0) (2019-09-10)